Refactor Drew main method.
Refactored UI, Storage, Parsing and List functionality into respective packages and classes.
