--- src/main/java/Command.java
@@ -1,3 +1,0 @@
-public enum Command {
-    LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, UNKNOWN
-}
