+++ src/main/java/drew/command/Command.java
@@ -1,0 +1,5 @@
+package drew.command;
+
+public enum Command {
+    LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, UNKNOWN, BYE
+}
