+++ src/main/java/drew/storage/TaskList.java
@@ -1,0 +1,20 @@
+package drew.storage;
+
+import drew.task.Task;
+
+import java.util.ArrayList;
+
+public class TaskList {
+    ArrayList<Task> tasks;
+
+    public TaskList() {
+        this.tasks = new ArrayList<Task>();
+    }
+    public TaskList(ArrayList<Task> ls) {
+        this.tasks = ls;
+    }
+
+    public ArrayList<Task> getList(){
+        return tasks;
+    }
+}
